backend/foodgram/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==2.2.16
django-cors-headers==3.11.0
django-extra-fields==3.0.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
isort==5.10.1
mccabe==0.6.1
pep8-naming==0.12.1
Pillow==9.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
testfixtures==6.18.3
toml==0.10.2
tzdata==2021.5
gunicorn==20.1.0
sqlparse==0.4.2
djoser==2.1.0
unicodecsv==0.14.1